requirements.txt
pygame
opencv-python
mediapipe
numpy
pillow
SpeechRecognition
PyAudio
